Add draft of Arduino Read SD/Write Lights loop
Lots of debugging in here.  Currently only handles one file on the SD
card - no logic yet to switch to the next when the file runs out.
Timing is correct, though: every 50 ms, one LED refresh of all 200
lights is sent, then a new frame is read and we busy loop for the
remainder of the 50ms time slice.