Add a splash screen
Before, a failure to read the SD card was indistinguishable from
a hardware or wiring problem with the lights, since all data ever sent
to the lights was first read from the SD card.

With this splash screen, we'll get a static image even if the SD card
can't be read.